chore: refactor BDD utilities to consolidate common code in lsadf-core
Move common BDD utilities from lsadf-api and lsadf-admin to lsadf-core module:
- Consolidate BddFieldConstants, BddUtils, and ParameterizedTypeReferenceUtils
- Update all BDD step definitions to import from com.lsadf.core.bdd
- Configure Maven test-jar plugin for proper test dependency sharing
- Remove duplicate code while maintaining module-specific functionality

This refactoring eliminates code duplication and follows best practices by
keeping common utilities in the core module while preserving module-specific
BDD step definitions in their respective modules.
