+++ lsadf_admin/src/test/java/com/lsadf/admin/application/bdd/given/BddGivenStepDefinitions.java
@@ -17,9 +17,9 @@ package com.lsadf.admin.application.bdd.given;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.lsadf.admin.application.bdd.BddFieldConstants;
+import com.lsadf.core.bdd.BddFieldConstants;
 import com.lsadf.admin.application.bdd.BddLoader;
-import com.lsadf.admin.application.bdd.BddUtils;
+import com.lsadf.core.bdd.BddUtils;
 import com.lsadf.admin.application.bdd.CacheEntryType;
 import com.lsadf.core.domain.game.save.characteristics.Characteristics;
 import com.lsadf.core.domain.game.save.currency.Currency;
