+++ lsadf_admin/src/test/java/com/lsadf/admin/application/bdd/then/BddCacheThenStepDefinitions.java
@@ -15,12 +15,12 @@
  */
 package com.lsadf.admin.application.bdd.then;
 
-import static com.lsadf.admin.application.bdd.BddFieldConstants.GameSave.USER_EMAIL;
-import static com.lsadf.admin.application.bdd.BddFieldConstants.StageCacheEntry.GAME_SAVE_ID;
+import static com.lsadf.core.bdd.BddFieldConstants.GameSave.USER_EMAIL;
+import static com.lsadf.core.bdd.BddFieldConstants.StageCacheEntry.GAME_SAVE_ID;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lsadf.admin.application.bdd.BddLoader;
-import com.lsadf.admin.application.bdd.BddUtils;
+import com.lsadf.core.bdd.BddUtils;
 import com.lsadf.admin.application.bdd.CacheEntryType;
 import com.lsadf.core.domain.game.save.characteristics.Characteristics;
 import com.lsadf.core.domain.game.save.currency.Currency;
