+++ lsadf_admin/src/test/java/com/lsadf/admin/application/bdd/when/BddCurrencyWhenStepDefinitions.java
@@ -18,7 +18,7 @@ package com.lsadf.admin.application.bdd.when;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lsadf.admin.application.bdd.BddLoader;
-import com.lsadf.admin.application.bdd.BddUtils;
+import com.lsadf.core.bdd.BddUtils;
 import com.lsadf.core.domain.game.save.currency.Currency;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.When;
