+++ lsadf_admin/src/test/java/com/lsadf/admin/application/bdd/when/BddStageWhenStepDefinitions.java
@@ -18,7 +18,7 @@ package com.lsadf.admin.application.bdd.when;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lsadf.admin.application.bdd.BddLoader;
-import com.lsadf.admin.application.bdd.BddUtils;
+import com.lsadf.core.bdd.BddUtils;
 import com.lsadf.core.domain.game.save.stage.Stage;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.When;
