+++ lsadf_api/src/test/java/com/lsadf/application/bdd/then/BddThenStepDefinitions.java
@@ -18,7 +18,7 @@ package com.lsadf.application.bdd.then;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lsadf.application.bdd.BddLoader;
-import com.lsadf.application.bdd.BddUtils;
+import com.lsadf.core.bdd.BddUtils;
 import com.lsadf.core.domain.game.save.GameSave;
 import com.lsadf.core.domain.game.save.characteristics.Characteristics;
 import com.lsadf.core.domain.game.save.currency.Currency;
